requirements.txt
click>=8.0.0
pyyaml>=6.0
docker>=6.0.0
gitpython>=3.1.0
jinja2>=3.0.0
cryptography>=3.4.0
jsonschema>=4.0.0